Rename import to external
- Since we will be exporting Procfiles and compose files, the package
name "import" is no longer appropriate. Renaming to "external" since it
is code that deals with external data formats.
