+++ cmd/compose2exo/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/deref/exo/internal/import/compose"
+	"github.com/deref/exo/internal/external/compose"
 	"github.com/deref/exo/internal/manifest"
 	"github.com/deref/exo/internal/util/cmdutil"
 )
