+++ cmd/procfile2exo/main.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/deref/exo/internal/import/procfile"
+	"github.com/deref/exo/internal/external/procfile"
 	"github.com/deref/exo/internal/manifest"
 	"github.com/deref/exo/internal/util/cmdutil"
 )
