+++ internal/core/server/config.go
@@ -10,8 +10,8 @@ import (
 	"strings"
 
 	"github.com/deref/exo/internal/core/api"
-	"github.com/deref/exo/internal/import/compose"
-	"github.com/deref/exo/internal/import/procfile"
+	"github.com/deref/exo/internal/external/compose"
+	"github.com/deref/exo/internal/external/procfile"
 	"github.com/deref/exo/internal/manifest"
 	"github.com/deref/exo/internal/util/errutil"
 	"github.com/deref/exo/internal/util/osutil"
