+++ internal/core/server/workspace.go
@@ -10,8 +10,8 @@ import (
 	"github.com/deref/exo/internal/chrono"
 	"github.com/deref/exo/internal/core/api"
 	state "github.com/deref/exo/internal/core/state/api"
+	"github.com/deref/exo/internal/external/procfile"
 	"github.com/deref/exo/internal/gensym"
-	"github.com/deref/exo/internal/import/procfile"
 	logd "github.com/deref/exo/internal/logd/api"
 	"github.com/deref/exo/internal/manifest"
 	"github.com/deref/exo/internal/providers/core"
